K - 9:36 pm fixed pycode style

+++ tests/test_models/test_base_model.py
@@ -22,9 +22,9 @@ class test_basemodel(unittest.TestCase):
         pass
 
     def tearDown(self):
-        try:
+        if os.path.exists(file.json):
             os.remove('file.json')
-        except:
+        else:
             pass
 
     def test_default(self):
